requirements.txt
Django
psycopg2-binary
django-countries
djangorestframework
djangorestframework-simplejwt